app/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.2
pydantic-settings==2.0.1
python-multipart==0.0.6
python-json-logger==2.0.7
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
prometheus-client==0.17.0
ultralytics==8.0.145  # Using a specific version that works
tritonclient[grpc]==2.31.0
torch==2.0.1
inference